Feat: ScheduledCard 컴포넌트, 유틸함수 분리
